refactor dockerfile, go modules with golang v1.14.2; "closes #163" - removed dep-related comands and notes - added two Dockerfiles (works with Podman, Buildah too)   - "prod": for clean build, with tests, runnng with non-root user, from scratch container   - "dev":  there is caching for golang deps, without tests, run as root, and result container can be not from scratch, for debug reasons. - docker-compose file:   - added shared context-build block, DRY.   - removed command name `jocko` in line `command: jocko broker...` because of we using the same ENTRYPOINT in docker file.
- small fixes in module paths:
  - "github.com/hashicorp/consul/testutil/retry" -> "github.com/hashicorp/consul/sdk/testutil/retry"

- update go.mod to the lastest version for all transitive deps (exepts github.com/mitchellh/go-testing-interface v1.4.0)
