+++ Makefile
@@ -4,8 +4,7 @@ DOCKER_TAG := latest
 all: test
 
 deps:
-	@which dep 2>/dev/null || go get -u github.com/golang/dep/cmd/dep
-	@dep ensure -v
+	@go mod download
 
 vet:
 	@go list ./... | grep -v vendor | xargs go vet
