+++ docker-compose.yml
@@ -1,20 +1,25 @@
 version: '3'
 
+x-context-build:
+  &context-build
+  context: .
+  dockerfile: Dockerfile-dev
+
 services:
   jocko_a:
     build:
-      context: .
+      <<: *context-build
     image: travisjeffery/jocko:latest
-    command: jocko broker --id 0 --raft-addr=jocko_a:9093 --bootstrap --bootstrap-expect=3
+    command: broker --id 0 --raft-addr=jocko_a:9093 --bootstrap --bootstrap-expect=3
 
   jocko_b:
     build:
-      context: .
+      <<: *context-build
     image: travisjeffery/jocko:latest
-    command: jocko broker --join=jocko_a:9094 --id 1 --raft-addr=jocko_b:9093 --bootstrap-expect=3
+    command: broker --join=jocko_a:9094 --id 1 --raft-addr=jocko_b:9093 --bootstrap-expect=3
 
   jocko_c:
     build:
-      context: .
+      <<: *context-build
     image: travisjeffery/jocko:latest
-    command: jocko broker --join=jocko_a:9094 --id 2 --raft-addr=jocko_c:9093 --bootstrap-expect=3
+    command: broker --join=jocko_a:9094 --id 2 --raft-addr=jocko_c:9093 --bootstrap-expect=3
